Function ataabu::get_float

pub(crate) fn get_float(
    input_str: String, 
    whole_token: &mut WholePositionToken
) -> (Option<Token>, String)

get_float token from a input string and return the float token and return a pair of (maybe, string) maybe that may be nothing or containing the result string is the remained token, typed "FLO".

float = [0-9]+[.][0-9]+

arguments

input_str - the string to be tokenized

return value

(Option<Token>, String)

the 1st Option<Token> : the result whose token.tyle="FLO" or None

the second String : the ramained surplus string

 
   let e = get_float("12.345 123454 as ".to_string(), &mut whole_token);
    // Some(Token { token_str: "12.345", token_type: "FLO", line_no: 1, col_no: 1 }) ,  123454 as
 
println!("{:?} , {}", e.0, e.1); 
   let f = get_float("123454 as ".to_string(), &mut whole_token)
    //None , 123454 as
    println!("{:?} , {}", f.0, f.1);