Function ataabu::match_range

pub(crate) fn match_range(
    char_min: String, 
    char_max: String, 
    maybe_string: ProcessedString
) -> ProcessedString

input the char as a String with min index, and the char as a String with max index, and the string processed by the lexer wrapped as ProcessedString, then return the value is: if it successed to be processed, the 1st char in remained is moved to the tail of the string matched. otherwise, the successed indicator ‘is_succeed’ is set to False. eg.

   let a =  ProcessedString {
   is_succeed: true,
   matched: "".to_string(),
   remained: "c59".to_string()};
    
   let b =  ProcessedString {
   is_succeed: true,
   matched: "".to_string(),
   remained: "159".to_string()};
    
   let a_res =  match_range("a".to_string(), "z".to_string(), a); // matched
   let b_res =  match_range("a".to_string(), "z".to_string(), b); // unmatched
    
   println!("succeed: {}, char: {}", a_res.is_succeed, a_res.matched);
   println!("succeed: {}, char: {}", b_res.is_succeed, b_res.matched);
 
// returns:
succeed: true, matched: c
succeed: false, matched: 

arguments

char_min : the character as the lower bound of index of a range of a character

char_max: the character as the upper bound of index of a range of a character

maybe_string: the container of the processed string.

return value

ProcessedString:

matched nothing => is_succeed attr. is false

match a character => is_succeed is true.